+++ app/adapters/node-storage.ts
@@ -1,10 +1,10 @@
 import OsfAdapter from './osf-adapter';
 
-export default class ProviderAdapter extends OsfAdapter {
+export default class NodeStorageAdapter extends OsfAdapter {
 }
 
 declare module 'ember-data/types/registries/adapter' {
     export default interface AdapterRegistry {
-        provider: ProviderAdapter;
+        'node-storage': NodeStorageAdapter;
     } // eslint-disable-line semi
 }