+++ app/models/contributor.ts
@@ -41,9 +41,11 @@ export default class ContributorModel extends OsfModel.extend(Validations) {
     @attr('boolean') bibliographic!: boolean;
     @attr('fixstring') unregisteredContributor?: string;
     @attr('number') index!: number;
+    @attr('string') sendEmail!: 'default' | 'preprint' | 'false';
+
+    // Write-only attributes
     @attr('fixstring') fullName!: string;
     @attr('fixstring') email!: string;
-    @attr('string') sendEmail!: 'default' | 'preprint' | 'false';
 
     @belongsTo('user', { inverse: 'contributors' })
     users!: DS.PromiseObject<UserModel> & UserModel;