+++ app/models/review-action.ts
@@ -1,5 +1,7 @@
 import { computed } from '@ember/object';
+import { inject as service } from '@ember/service';
 import DS from 'ember-data';
+import Intl from 'ember-intl/services/intl';
 
 import OsfModel from './osf-model';
 import RegistrationModel from './registration';
@@ -9,38 +11,27 @@ import UserModel from './user';
 const { attr, belongsTo } = DS;
 
 export enum ReviewActionTrigger {
-    Accept = 'accept', // accept submission
-    Reject = 'reject', // reject submission
-    Embargo = 'embargo', // embargo (accept) submission
+    AcceptSubmission = 'accept_submission', // accept submission
+    RejectSubmission = 'reject_submission', // reject submission
     ForceWithdraw = 'force_withdraw', // force withdraw without request
-    RequestWithdraw = 'request_withdraw', // request to withdraw by contributors
-    WithdrawRequestPasses = 'withdraw_request_passes', // other contributors approves requests
-    WithdrawRequestFails = 'withdraw_request_fails', // other contributors disapprove of requests
-    Withdraw = 'withdraw', // accept withdrawal request
-    RejectWithdraw = 'reject_withdraw', // deny withdrawal request
-    RequestEmbargoTermination = 'request_embargo_termination', // request to end embargo early
-    TerminateEmbargo = 'terminate_embargo', // contributors approve request to end embargo early
+    RequestWithdrawal = 'request_withdrawal', // request to withdraw by contributors
+    AcceptWithdrawal = 'accept_withdrawal', // accept withdrawal request
+    RejectWithdrawal = 'reject_withdrawal', // deny withdrawal request
 }
 
-// make sure these match the reviewActionTriggers
-const TriggersPastTenseMap: Record<string, string> = {
-    submit: 'Submitted',
-    accept: 'Accepted',
-    reject: 'Rejected',
-    edit_comment: 'Comment edited',
-    embargo: 'Embargoed',
-    withdraw: 'Withdrawn',
-    request_withdraw: 'Withdrawal requested',
-    withdraw_request_passes: 'Withdrawal request accepted',
-    withdraw_request_fails: 'Withdrawal rejected',
-    force_withdraw: 'Withdrawal forced',
-    request_embargo: 'Embargo requested',
-    request_embargo_termination: 'Embargo termination requested',
-    terminate_embargo: 'Embargo terminated',
+const TriggerToPastTenseTranslationKey: Record<ReviewActionTrigger, string> = {
+    accept_submission: 'registries.reviewActions.triggerPastTense.accept_submission',
+    reject_submission: 'registries.reviewActions.triggerPastTense.reject_submission',
+    force_withdraw: 'registries.reviewActions.triggerPastTense.force_withdraw',
+    request_withdrawal: 'registries.reviewActions.triggerPastTense.request_withdrawal',
+    accept_withdrawal: 'registries.reviewActions.triggerPastTense.accept_withdrawal',
+    reject_withdrawal: 'registries.reviewActions.triggerPastTense.reject_withdrawal',
 };
 
 export default class ReviewActionModel extends OsfModel {
-    @attr('string') actionTrigger!: string;
+    @service intl!: Intl;
+
+    @attr('string') actionTrigger!: ReviewActionTrigger;
     @attr('string') comment!: string;
     @attr('string') fromState!: string;
     @attr('string') toState!: string;
@@ -60,7 +51,8 @@ export default class ReviewActionModel extends OsfModel {
 
     @computed('actionTrigger')
     get triggerPastTense(): string {
-        return TriggersPastTenseMap[this.actionTrigger];
+        const translationKey = TriggerToPastTenseTranslationKey[this.actionTrigger];
+        return translationKey ? this.intl.t(translationKey) : '';
     }
 }