+++ lib/app-components/addon/helpers/get-model.ts
@@ -1,0 +1,13 @@
+import Helper from '@ember/component/helper';
+import ObjectProxy from '@ember/object/proxy';
+import DS from 'ember-data';
+import OsfModel from 'ember-osf-web/models/osf-model';
+
+export default class GetModelHelper extends Helper {
+    compute([model]: [DS.PromiseObject<OsfModel> | OsfModel]) {
+        if (!model) {
+            return undefined;
+        }
+        return model instanceof ObjectProxy ? model.content : model;
+    }
+}