+++ mirage/config.ts
@@ -1,6 +1,7 @@
 import { Server } from 'ember-cli-mirage';
 import config from 'ember-get-config';
 
+import { createReviewAction } from 'ember-osf-web/mirage/views/review-action';
 import { addModerator } from './views/addModerator';
 import { getCitation } from './views/citation';
 import { searchCollections } from './views/collection-search';
@@ -19,7 +20,7 @@ import { createFork, createRegistrationFork } from './views/fork';
 import { guidDetail } from './views/guid';
 import { identifierCreate } from './views/identifier';
 import { summaryMetrics } from './views/institution';
-import { createNode } from './views/node';
+import { createNode, storageStatus } from './views/node';
 import { osfNestedResource, osfResource, osfToManyRelationship } from './views/osf-resource';
 import { getProviderSubjects } from './views/provider-subjects';
 import { createRegistration, forkRegistration, registrationDetail } from './views/registration';
@@ -109,6 +110,7 @@ export default function(this: Server) {
         only: ['related', 'add', 'remove'],
         path: '/nodes/:parentID/relationships/institutions',
     });
+    this.get('/nodes/:id/storage', storageStatus);
 
     osfToManyRelationship(this, 'node', 'subjects', {
         only: ['related', 'self'],
@@ -137,6 +139,7 @@ export default function(this: Server) {
         path: '/registrations/:parentID/review-actions',
         relatedModelName: 'review-action',
     });
+    this.post('/registrations/:parentID/review-actions', createReviewAction);
     this.post('/registrations/:id/forks', forkRegistration);
 
     osfNestedResource(this, 'registration', 'contributors', { defaultSortKey: 'index' });