+++ mirage/views/node.ts
@@ -1,4 +1,5 @@
-import { HandlerContext, Schema } from 'ember-cli-mirage';
+import { HandlerContext, Request, Schema } from 'ember-cli-mirage';
+import { process } from './utils';
 
 export function createNode(this: HandlerContext, schema: Schema) {
     const attrs = this.normalizedRequestAttrs('node');
@@ -17,4 +18,10 @@ export function createNode(this: HandlerContext, schema: Schema) {
     }
 
     return node;
+}
+
+export function storageStatus(this: HandlerContext, schema: Schema, request: Request) {
+    const model = this.serialize(schema.storage.find(request.params.id)).data;
+    const data = process(schema, request, this, [model]).data[0];
+    return { data };
 }